coll/base: silence misc warning
as reported by Coverity with CIDs 1363349-1363362
(back-ported from commit open-mpi/ompi@7b8094aac12b33119e327aa0692fcd8d0e85a958)

+++ ompi/mca/coll/basic/coll_basic_allgather.c
@@ -158,12 +158,11 @@ mca_coll_basic_allgather_inter(void *sbuf, int scount,
             goto exit;
         }
 
-        span = opal_datatype_span(&sdtype->super, scount * size, &gap);
+        /* Step 2: exchange the resuts between the root processes */
+        span = opal_datatype_span(&sdtype->super, (int64_t)scount * (int64_t)size, &gap);
         tmpbuf_free = (char *) malloc(span);
-        if (NULL == tmpbuf_free) {
-            err = OMPI_ERR_OUT_OF_RESOURCE;
-            goto exit;
-        }
+        if (NULL == tmpbuf_free) { err = OMPI_ERR_OUT_OF_RESOURCE; goto exit; }
+
         tmpbuf = tmpbuf_free - gap;
 
         err = MCA_PML_CALL(isend(rbuf, rsize * rcount, rdtype, 0,
